[css-text] Keep track of the initial offset for each handled text
In order to identify single leading white-spaces as potential line
breaking opportunities we need to know the offset the iterator starts
at for the current text to handle.

When we landed te change in r596406 we incorrectly assumed that leading
white-spaces were always at offset 1. However, this is not valid when
such leading white-spaces are generated by previous broken lines.

This CL solves the issue by using the stored initial offset to identify
such single leading white-spaces as potential breaking opportunities.

Bug: 900727
Change-Id: Ibc9dda7632d72ed82a61025015489ec3b70d6919